Add function for finding common interfaces
